initial in c++ instead of python (2048x2048 ~3 seconds -> 0.45)

+++ src/backend/gpu_backend.metal
@@ -1,8 +1,10 @@
 #include <metal_stdlib>
 using namespace metal;
+//#include "tensor.h"
 
-kernel void simple_shader(device float* input [[ buffer(0) ]],
-                          device float* output [[ buffer(1) ]],
+// half is a special data type that represents floats in 16 bits
+kernel void simple_shader(device half* input [[ buffer(0) ]],
+                          device half* output [[ buffer(1) ]],
                           uint id [[ thread_position_in_grid ]]) {
     output[id] = input[id] * 2.0;
 }
